pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["knockoph <knockoph@proton.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
